requirements_tr_etp.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
tabula-py==2.8.2
pandas==2.1.3
google-generativeai==0.3.2
